Version 0.8.1 release. More bot api support, methods for getting an id from a name.
Overhauls access to users, groups, and channels, as well as their ids.
Adds support for setting the topic and purpose of a group, posting
messages via the web api, adding reactions to messages, files, and file
comments, as well as updating the list of users, groups, and channels
via the web api.