Add an async message send API
The current interface only allows for sending messages synchronously in
response to some event. In the past, it was possible to get an
mpsc::Sender<WsMessage> to deliver messages asynchronously. This commit
restores that functionality with a slighly more ergonomic API.

The RtmClient has a new method

    pub fn channel(&self) -> Option<Sender>

which returns a slack::Sender if the RtmClient has been logged in
(creating the slack::Sender requires the RtmClient's mpsc::Sender to
initialize itself).

`Sender` has a high and low level send API. The high level API

    pub fn send_message_chid(&self, channel_id: &str, text: &str)

will automatically increment the message id, build the JSON payload, and
send it to the event loop for delivery. The low level API is two parts

    pub fn get_msg_uid(&self) -> isize

for getting the next message id, and

    pub fn send(&self, raw: &str)

for sending a pre-built JSON payload.