Retry receive message on EAGAIN
RtmClient.run() currently will return an Error if any error is
encountered. This is generally what we want, but not always. Recovering
from this error previously meant having to login again and recreate
async sender. Since this error is recoverable internally, we should try
to so clients don't need to do the reconnect dance.